backend/requirements.txt
django==5.2.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
pymongo==4.15.4
python-dotenv==1.2.1
pytz==2025.2
mongoengine==0.29.1
djangorestframework-simplejwt==5.5.1
